Check if MEX dir exists before removing it when an exception occurs

+++ gemma-core/src/main/java/ubic/gemma/core/analysis/service/ExpressionDataFileServiceImpl.java
@@ -625,7 +625,9 @@ public class ExpressionDataFileServiceImpl implements ExpressionDataFileService 
                 return writer.write( matrix, cs2gene, destDir );
             }
         } catch ( Exception e ) {
-            PathUtils.deleteDirectory( destDir );
+            if ( Files.exists( destDir ) ) {
+                PathUtils.deleteDirectory( destDir );
+            }
             throw e;
         }
     }
